docs(Divider): move to component folder
