+++ src/components/Divider/docs/DividerExtendedExample.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {Divider} from '../../src/components/Divider';
+import {Divider} from '../Divider';
 import extendedStyles from './DividerExtendedExample.st.css';
 
 export const DividerExtendedExample = (props) => <Divider {...props} {...extendedStyles('root', {}, props)}/>;
