+++ src/components/Divider/docs/examples.tsx
@@ -4,7 +4,8 @@ import CodeExample from 'wix-storybook-utils/CodeExample';
 
 import * as ExtendedRawSource from '!raw-loader!./DividerExtendedExample.tsx';
 import * as ExtendedCSSRawSource from '!raw-loader!./DividerExtendedExample.st.css';
-import {MockSettings} from '../helperComponents/MockSettings';
+
+import {MockSettings} from '../../../../stories/helperComponents/MockSettings';
 import {DividerExtendedExample} from './DividerExtendedExample';
 
 export const Examples = () => (
